Fixed bug when using Mob Rules
The bug was caused by a change implemented for rolling individually, which I forgot to double check for the case where you're using Mob Rules instead.

Another change addresses the incorrectly configured ammo problem from releases 0.1.14-0.1.16. It now not only gives an error message, but then skips the attack and continues with the next attack.